+++ lib/ts/utils/index.ts
@@ -116,7 +116,7 @@ export function validateAndNormaliseInputOrThrowError(options: InputType): Norma
         sessionTokenBackendDomain = normaliseSessionScopeOrThrowError(options.sessionTokenBackendDomain);
     }
 
-    let maxRetryAttemptsForSessionRefresh = 3;
+    let maxRetryAttemptsForSessionRefresh = 10;
     if (options.maxRetryAttemptsForSessionRefresh !== undefined) {
         if (options.maxRetryAttemptsForSessionRefresh < 0) {
             throw new Error("maxRetryAttemptsForSessionRefresh must be greater than or equal to 0.");